Add "clientRepository"

+++ src/main/java/com/sbn/app/entity/Client.java
@@ -40,7 +40,7 @@ public class Client implements Serializable {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Basic(optional = false)
     @Column(name = "id")
-    private Integer id;
+    private Long id;
     @Size(max = 100)
     @Column(name = "name")
     private String name;
@@ -62,15 +62,15 @@ public class Client implements Serializable {
     public Client() {
     }
 
-    public Client(Integer id) {
+    public Client(Long id) {
         this.id = id;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
